feat(image)!: [MacOS] Make Image inherit from FrameworkElement
BREAKING CHANGE: Image now inherits correctly from FrameworkElement, instead of directly from the native type. The native image is now internally set as a child of the Windows.UI.Xaml.Controls.Image.
